+++ HISTORY.md
@@ -57,6 +57,7 @@
 * Objects, Threads: lazily initialize volume property
 * Objects: use AudioContext to play recorded sounds
 * Objects: new audio scheme support for the stage
+* Objects: added basic stereo-panning support for sounds (under construction)
 
 ### 2019-04-01
 * Objects: let the Microphone share the Note prototype's AudioContext